+++ src/atlas/mpl/GatherScatter.cc
@@ -136,7 +136,7 @@ void GatherScatter::setup( const int part[],
   glbcnt_ = std::accumulate(glbcounts_.begin(),glbcounts_.end(),0);
 
   glbdispls_[0]=0;
-  for (size_t jproc=1; jproc<nproc; ++jproc) // start at 1
+  for (int jproc = 1; jproc < nproc; ++jproc) // start at 1
   {
     glbdispls_[jproc]=glbcounts_[jproc-1]+glbdispls_[jproc-1];
   }
@@ -169,7 +169,7 @@ void GatherScatter::setup( const int part[],
   }
   glbdispls_[0]=0;
 
-  for (size_t jproc=1; jproc<nproc; ++jproc) // start at 1
+  for (int jproc = 1; jproc < nproc; ++jproc) // start at 1
   {
     glbdispls_[jproc]=glbcounts_[jproc-1]+glbdispls_[jproc-1];
   }
@@ -220,7 +220,7 @@ void GatherScatter::setup( const int part[],
   {
     IsGhostPoint is_ghost(part,remote_idx,base,parsize_);
     maxgid = -1;
-    for (int jj=0; jj<parsize_; ++jj)
+    for (size_t jj = 0; jj < parsize_; ++jj)
     {
       if ( !is_ghost(jj) || include_ghost )
       {
@@ -260,7 +260,7 @@ void GatherScatter::setup( const int part[],
   glbcnt_ = std::accumulate(glbcounts_.begin(),glbcounts_.end(),0);
 
   glbdispls_[0]=0;
-  for (size_t jproc=1; jproc<nproc; ++jproc) // start at 1
+  for (int jproc = 1; jproc < nproc; ++jproc) // start at 1
   {
     glbdispls_[jproc]=glbcounts_[jproc-1]+glbdispls_[jproc-1];
   }
@@ -270,7 +270,7 @@ void GatherScatter::setup( const int part[],
                       root_, eckit::mpi::comm()) );
 
   // Load recvnodes in sorting structure
-  int nb_recv_nodes = glbcnt_/3;
+  size_t nb_recv_nodes = glbcnt_/3;
   std::vector<Node> node_sort(nb_recv_nodes);
   nodes = ArrayView<int,2> (recvnodes.data(),make_shape(nb_recv_nodes,3));
   for( size_t n=0; n<nb_recv_nodes; ++n )
@@ -314,13 +314,13 @@ void GatherScatter::setup( const int part[],
   // Assemble list to ask needed
   glbcounts_.assign(nproc,0);
   glbdispls_.assign(nproc,0);
-  for( int n=0; n<node_sort.size(); ++n )
+  for(size_t n = 0; n < node_sort.size(); ++n)
   {
     ++glbcounts_[node_sort[n].p] ;
   }
   glbdispls_[0]=0;
 
-  for (size_t jproc=1; jproc<nproc; ++jproc) // start at 1
+  for (int jproc = 1; jproc < nproc; ++jproc) // start at 1
   {
     glbdispls_[jproc]=glbcounts_[jproc-1]+glbdispls_[jproc-1];
   }
@@ -404,7 +404,7 @@ void atlas__GatherScatter__setup32 ( GatherScatter* This,  int part[],
   This->setup(part,remote_idx,base,glb_idx,max_glb_idx,parsize);
 #else
   std::vector<gidx_t> glb_idx_convert(parsize);
-  for( size_t j=0; j<parsize; ++j )
+  for(int j = 0; j < parsize; ++j)
   {
     glb_idx_convert[j] = glb_idx[j];
   }
@@ -442,11 +442,11 @@ void atlas__GatherScatter__gather_int ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -462,11 +462,11 @@ void atlas__GatherScatter__gather_long ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -481,11 +481,11 @@ void atlas__GatherScatter__gather_float ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -500,11 +500,11 @@ void atlas__GatherScatter__gather_double ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -519,11 +519,11 @@ void atlas__GatherScatter__scatter_int ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -539,11 +539,11 @@ void atlas__GatherScatter__scatter_long ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -558,11 +558,11 @@ void atlas__GatherScatter__scatter_float ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
@@ -577,11 +577,11 @@ void atlas__GatherScatter__scatter_double ( GatherScatter* This,
   std::vector<size_t> lvextents(lvar_rank);
   std::vector<size_t> gvstrides(gvar_rank);
   std::vector<size_t> gvextents(gvar_rank);
-  for( size_t n=0; n<lvar_rank; ++n ) {
+  for(int n = 0; n < lvar_rank; ++n) {
     lvstrides[n] = lvar_strides[n];
     lvextents[n] = lvar_extents[n];
   }
-  for( size_t n=0; n<gvar_rank; ++n ) {
+  for(int n = 0; n < gvar_rank; ++n) {
     gvstrides[n] = gvar_strides[n];
     gvextents[n] = gvar_extents[n];
   }
